Refactor Leaderboard logic into dedicated manager
